+++ src/utils/constructionUtil.ts
@@ -77,10 +77,7 @@ export async function addConstructionUtil(
     }
     //save model if images and audio files are uploaded
     if (
-      attachmentsIds.length > 0 &&
-      audioPath !== undefined &&
-      audioPath !== "" &&
-      audioPath !== null
+      attachmentsIds.length > 0 
     ) {
       model.attachmentsPath = attachmentsIds;
       const newModel = new Construction(model);
@@ -95,7 +92,9 @@ export async function addConstructionUtil(
         };
         return response;
       } else {
-        fs.unlinkSync(audioPath);
+        if(audioPath && audioPath !== ''){
+          fs.unlinkSync(audioPath)
+        }
         for (let k = 0; k < path.length; k++) {
           fs.unlinkSync(path[k]);
         }
